@@ -14,18 +14,18 @@ use App\Services\BidService;
 use App\Services\FareService;
 use App\Services\FlightService;
 
-class BidTest extends TestCase
+final class BidTest extends TestCase
 {
     /** @var BidService */
-    protected $bidSvc;
+    protected BidService $bidSvc;
 
     /** @var FlightService */
-    protected $flightSvc;
+    protected FlightService $flightSvc;
 
     /** @var SettingRepository */
-    protected $settingsRepo;
+    protected SettingRepository $settingsRepo;
 
-    public function setUp(): void
+    protected function setUp(): void
     {
         parent::setUp();
         $this->addData('base');
@@ -35,30 +35,12 @@ class BidTest extends TestCase
         $this->settingsRepo = app(SettingRepository::class);
     }
 
-    public function addFlight($user, $subfleet = null)
-    {
-        $flight = Flight::factory()->create([
-            'airline_id' => $user->airline_id,
-        ]);
-
-        if ($subfleet === null) {
-            /** @var Subfleet $subfleet */
-            $subfleet = Subfleet::factory()->create([
-                'airline_id' => $user->airline_id,
-            ])->id;
-        }
-
-        $flight->subfleets()->syncWithoutDetaching([$subfleet]);
-
-        return $flight;
-    }
-
     /**
      * Add/remove a bid, test the API, etc
      *
-     * @throws Exception|\Exception
+     * @throws \Exception
      */
-    public function testBids()
+    public function testBids(): void
     {
         $this->settingsRepo->store('bids.allow_multiple_bids', true);
         $this->settingsRepo->store('bids.disable_flight_on_bid', false);
@@ -85,7 +67,7 @@ class BidTest extends TestCase
         $headers = $this->headers($user);
 
         /** @var Flight $flight */
-        $flight = $this->addFlight($user, $subfleet['subfleet']->id);
+        $flight = $this->addFlight($user, [], $subfleet['subfleet']->id);
 
         $bid = $this->bidSvc->addBid($flight, $user, $subfleet['aircraft'][0]);
         $this->assertEquals($user->id, $bid->user_id);
@@ -163,7 +145,7 @@ class BidTest extends TestCase
         $this->assertCount(0, $body);
     }
 
-    public function testMultipleBidsSingleFlight()
+    public function testMultipleBidsSingleFlight(): void
     {
         $this->settingsRepo->store('bids.disable_flight_on_bid', true);
 
@@ -185,7 +167,7 @@ class BidTest extends TestCase
     /**
      * Add a flight bid VIA the API
      */
-    public function testAddBidApi()
+    public function testAddBidApi(): void
     {
         $this->user = User::factory()->create();
         $user2 = User::factory()->create();
@@ -222,8 +204,10 @@ class BidTest extends TestCase
 
     /**
      * Delete a flight and make sure all the bids are gone
+     *
+     * @throws \Exception
      */
-    public function testDeleteFlightWithBids()
+    public function testDeleteFlightWithBids(): void
     {
         $user = User::factory()->create();
         $headers = $this->headers($user);
@@ -262,8 +246,10 @@ class BidTest extends TestCase
 
     /**
      * Create a bid with an aircraft and check the aircraft booking
+     *
+     * @throws \Exception
      */
-    public function testBidWithAircraft()
+    public function testBidWithAircraft(): void
     {
         $this->settingsRepo->store('pireps.restrict_aircraft_to_rank', false);
         $this->settingsRepo->store('pireps.only_aircraft_at_dpt_airport', false);
@@ -277,7 +263,7 @@ class BidTest extends TestCase
         $subfleet = $this->createSubfleetWithAircraft(10);
         $aircraft = $subfleet['aircraft']->first();
 
-        $flight = $this->addFlight($user, $subfleet['subfleet']->id);
+        $flight = $this->addFlight($user, [], $subfleet['subfleet']->id);
         $flight->subfleets()->syncWithoutDetaching([$subfleet_unused['subfleet']->id]);
 
         $bid = $this->bidSvc->addBid($flight, $user, $aircraft);
@@ -293,7 +279,7 @@ class BidTest extends TestCase
         $this->assertEquals(1, $aircraft->bid->count());
 
         // Now add another bid on another flight with the same aircraft, should throw an exception
-        $flight2 = $this->addFlight($user, $subfleet['subfleet']->id);
+        $flight2 = $this->addFlight($user, [], $subfleet['subfleet']->id);
 
         $this->expectException(BidExistsForAircraft::class);
         $bid2 = $this->bidSvc->addBid($flight2, $user, $aircraft);
