@@ -27,27 +27,27 @@ use App\Support\Math;
 use App\Support\Money;
 use Exception;
 
-class FinanceTest extends TestCase
+final class FinanceTest extends TestCase
 {
-    /** @var \App\Repositories\ExpenseRepository */
-    private $expenseRepo;
+    /** @var ExpenseRepository */
+    private ExpenseRepository $expenseRepo;
 
-    /** @var \App\Services\FareService */
-    private $fareSvc;
+    /** @var FareService */
+    private FareService $fareSvc;
 
     /** @var PirepFinanceService */
-    private $financeSvc;
+    private PirepFinanceService $financeSvc;
 
     /** @var FleetService */
-    private $fleetSvc;
+    private FleetService $fleetSvc;
 
     /** @var PirepService */
-    private $pirepSvc;
+    private PirepService $pirepSvc;
 
     /**
      * @throws Exception
      */
-    public function setUp(): void
+    protected function setUp(): void
     {
         parent::setUp();
         $this->addData('base');
@@ -68,7 +68,7 @@ class FinanceTest extends TestCase
      *
      * @return array
      */
-    public function createFullPirep()
+    public function createFullPirep(): array
     {
         /**
          * Setup tests
@@ -187,7 +187,7 @@ class FinanceTest extends TestCase
      * The fares, etc for a subfleet has to be adjusted to the fleet
      * https://github.com/nabeelio/phpvms/issues/905
      */
-    public function testFlightFaresOverAPI()
+    public function testFlightFaresOverAPI(): void
     {
         $this->updateSetting('pireps.only_aircraft_at_dpt_airport', false);
         $this->updateSetting('pireps.restrict_aircraft_to_rank', false);
@@ -257,7 +257,7 @@ class FinanceTest extends TestCase
         $this->assertEquals(20, $body['subfleets'][0]['fares'][0]['cost']);
     }
 
-    public function testFlightFaresOverAPIOnUserBids()
+    public function testFlightFaresOverAPIOnUserBids(): void
     {
         $this->updateSetting('pireps.only_aircraft_at_dpt_airport', false);
         $this->updateSetting('pireps.restrict_aircraft_to_rank', false);
@@ -295,7 +295,7 @@ class FinanceTest extends TestCase
         $this->assertEquals($fare->capacity, $body[0]['flight']['fares'][0]['capacity']);
     }
 
-    public function testSubfleetFaresOverAPI()
+    public function testSubfleetFaresOverAPI(): void
     {
         $this->updateSetting('pireps.only_aircraft_at_dpt_airport', false);
         $this->updateSetting('pireps.restrict_aircraft_to_rank', false);
@@ -327,7 +327,7 @@ class FinanceTest extends TestCase
     /**
      * Assign percentage values and make sure they're valid
      */
-    public function testFlightFareOverrideAsPercent()
+    public function testFlightFareOverrideAsPercent(): void
     {
         /** @var Flight $flight */
         $flight = Flight::factory()->create();
@@ -362,7 +362,7 @@ class FinanceTest extends TestCase
         $this->assertEquals($new_capacity, $ac_fares[0]->capacity);
     }
 
-    public function testSubfleetFaresNoOverride()
+    public function testSubfleetFaresNoOverride(): void
     {
         $subfleet = Subfleet::factory()->create();
         $fare = Fare::factory()->create();
@@ -393,7 +393,7 @@ class FinanceTest extends TestCase
         $this->assertCount(0, $this->fareSvc->getForSubfleet($subfleet));
     }
 
-    public function testSubfleetFaresOverride()
+    public function testSubfleetFaresOverride(): void
     {
         $subfleet = Subfleet::factory()->create();
         $fare = Fare::factory()->create();
@@ -430,7 +430,7 @@ class FinanceTest extends TestCase
     /**
      * Assign percentage values and make sure they're valid
      */
-    public function testSubfleetFareOverrideAsPercent()
+    public function testSubfleetFareOverrideAsPercent(): void
     {
         $subfleet = Subfleet::factory()->create();
         $fare = Fare::factory()->create();
@@ -461,7 +461,7 @@ class FinanceTest extends TestCase
      * Test getting the fares from the flight svc. Have a few base fares
      * and then override some of them
      */
-    public function testGetFaresWithOverrides()
+    public function testGetFaresWithOverrides(): void
     {
         $flight = Flight::factory()->create();
         $subfleet = Subfleet::factory()->create();
@@ -515,7 +515,7 @@ class FinanceTest extends TestCase
         }
     }
 
-    public function testGetFaresNoFlightOverrides()
+    public function testGetFaresNoFlightOverrides(): void
     {
         $subfleet = Subfleet::factory()->create();
         [$fare1, $fare2, $fare3] = Fare::factory()->count(3)->create();
@@ -558,8 +558,10 @@ class FinanceTest extends TestCase
 
     /**
      * Get the pilot pay, derived from the rank
+     *
+     * @throws Exception
      */
-    public function testGetPilotPayNoOverride()
+    public function testGetPilotPayNoOverride(): void
     {
         $subfleet = $this->createSubfleetWithAircraft(2);
         $rank = $this->createRank(10, [$subfleet['subfleet']->id]);
@@ -581,8 +583,10 @@ class FinanceTest extends TestCase
 
     /**
      * Get the pilot pay, but include different overrides
+     *
+     * @throws Exception
      */
-    public function testGetPilotPayWithOverride()
+    public function testGetPilotPayWithOverride(): void
     {
         $acars_pay_rate = 100;
 
@@ -632,8 +636,10 @@ class FinanceTest extends TestCase
 
     /**
      * Get the payment for a pilot
+     *
+     * @throws Exception
      */
-    public function testGetPirepPilotPay()
+    public function testGetPirepPilotPay(): void
     {
         $acars_pay_rate = 100;
 
@@ -668,7 +674,10 @@ class FinanceTest extends TestCase
         $this->assertEquals($payment->getValue(), 150);
     }
 
-    public function testGetPirepPilotPayWithFixedPrice()
+    /**
+     * @throws Exception
+     */
+    public function testGetPirepPilotPayWithFixedPrice(): void
     {
         $acars_pay_rate = 100;
 
@@ -766,7 +775,7 @@ class FinanceTest extends TestCase
     /**
      * @throws Exception
      */
-    public function testPirepFares()
+    public function testPirepFares(): void
     {
         [$user, $pirep, $fares] = $this->createFullPirep();
 
@@ -796,7 +805,7 @@ class FinanceTest extends TestCase
     /**
      * Test that all expenses are pulled properly
      */
-    public function testPirepExpenses()
+    public function testPirepExpenses(): void
     {
         /** @var Airline $airline */
         $airline = Airline::factory()->create();
@@ -858,7 +867,7 @@ class FinanceTest extends TestCase
         $this->assertEquals($obj->id, $expense->ref_model_id);
     }
 
-    public function testAirportExpenses()
+    public function testAirportExpenses(): void
     {
         /** @var Airport $apt1 */
         $apt1 = Airport::factory()->create();
@@ -899,7 +908,7 @@ class FinanceTest extends TestCase
     /**
      * @throws Exception
      */
-    public function testPirepFinances()
+    public function testPirepFinances(): void
     {
         $journalRepo = app(JournalRepository::class);
 
@@ -956,7 +965,7 @@ class FinanceTest extends TestCase
     /**
      * @throws Exception
      */
-    public function testPirepFinancesSpecificExpense()
+    public function testPirepFinancesSpecificExpense(): void
     {
         $journalRepo = app(JournalRepository::class);
 
@@ -1078,7 +1087,7 @@ class FinanceTest extends TestCase
     /**
      * @throws Exception
      */
-    public function testPirepFinancesExpensesMultiAirline()
+    public function testPirepFinancesExpensesMultiAirline(): void
     {
         /** @var Airline $airline */
         $airline = Airline::factory()->create();
