@@ -29,15 +29,15 @@ use App\Support\Units\Fuel;
 use Carbon\Carbon;
 use Illuminate\Support\Facades\Notification;
 
-class PIREPTest extends TestCase
+final class PIREPTest extends TestCase
 {
     /** @var PirepService */
-    protected $pirepSvc;
+    protected PirepService $pirepSvc;
 
     /** @var SettingRepository */
-    protected $settingsRepo;
+    protected SettingRepository $settingsRepo;
 
-    public function setUp(): void
+    protected function setUp(): void
     {
         parent::setUp();
         $this->addData('base');
@@ -47,7 +47,7 @@ class PIREPTest extends TestCase
         $this->settingsRepo = app(SettingRepository::class);
     }
 
-    protected function createNewRoute()
+    protected function createNewRoute(): array
     {
         $route = [];
         $navpoints = Navdata::factory()->count(5)->create();
@@ -58,7 +58,7 @@ class PIREPTest extends TestCase
         return $route;
     }
 
-    protected function getAcarsRoute($pirep)
+    protected function getAcarsRoute(Pirep $pirep): array
     {
         $saved_route = [];
         $route_points = Acars::where(
@@ -73,9 +73,9 @@ class PIREPTest extends TestCase
     }
 
     /**
-     * @throws Exception
+     * @throws \Exception
      */
-    public function testAddPirep()
+    public function testAddPirep(): void
     {
         $user = User::factory()->create();
 
@@ -89,14 +89,14 @@ class PIREPTest extends TestCase
 
         try {
             $this->pirepSvc->saveRoute($pirep);
-        } catch (Exception $e) {
+        } catch (\Exception $e) {
             throw $e;
         }
 
         /*
          * Check the initial state info
          */
-        $this->assertEquals($pirep->state, PirepState::PENDING);
+        $this->assertEquals(PirepState::PENDING, $pirep->state);
 
         /**
          * Now set the PIREP state to ACCEPTED
@@ -165,8 +165,10 @@ class PIREPTest extends TestCase
 
     /**
      * Make sure the unit conversions look to be proper
+     *
+     * @throws \Exception
      */
-    public function testUnitFields()
+    public function testUnitFields(): void
     {
         /** @var Pirep $pirep */
         $pirep = $this->createPirep();
@@ -234,7 +236,7 @@ class PIREPTest extends TestCase
         );
     }
 
-    public function testGetUserPireps()
+    public function testGetUserPireps(): void
     {
         $this->user = User::factory()->create();
         $pirep_done = Pirep::factory()->create([
@@ -278,7 +280,7 @@ class PIREPTest extends TestCase
      *
      * @throws \Exception
      */
-    public function testPirepNotifications()
+    public function testPirepNotifications(): void
     {
         /** @var User $user */
         $user = User::factory()->create([
@@ -305,8 +307,10 @@ class PIREPTest extends TestCase
 
     /**
      * check the stats/ranks, etc have incremented properly
+     *
+     * @throws \Exception
      */
-    public function testPilotStatsIncr()
+    public function testPilotStatsIncr(): void
     {
         $this->updateSetting('pilots.count_transfer_hours', false);
 
@@ -382,7 +386,7 @@ class PIREPTest extends TestCase
      *
      * @throws \Exception
      */
-    public function testPilotDontChangeRank()
+    public function testPilotDontChangeRank(): void
     {
         /** @var Rank $rank */
         $rank = Rank::factory()->create([
@@ -418,8 +422,10 @@ class PIREPTest extends TestCase
 
     /**
      * check the stats/ranks, etc have incremented properly
+     *
+     * @throws \Exception
      */
-    public function testPilotStatsIncrWithTransferHours()
+    public function testPilotStatsIncrWithTransferHours(): void
     {
         $this->updateSetting('pilots.count_transfer_hours', true);
 
@@ -472,7 +478,7 @@ class PIREPTest extends TestCase
      * When a PIREP is filed by a user on leave, make sure they flip from leave to active
      * It doesn't matter if the PIREP is accepted or rejected
      */
-    public function testPilotStatusChange()
+    public function testPilotStatusChange(): void
     {
         /** @var \App\Models\User $user */
         $user = User::factory()->create([
@@ -498,7 +504,7 @@ class PIREPTest extends TestCase
     /**
      * Find and check for any duplicate PIREPs by a user
      */
-    public function testDuplicatePireps()
+    public function testDuplicatePireps(): void
     {
         $user = User::factory()->create();
         $pirep = Pirep::factory()->create([
@@ -523,7 +529,10 @@ class PIREPTest extends TestCase
         $this->assertFalse($dupe_pirep);
     }
 
-    public function testCancelViaAPI()
+    /**
+     * @throws \Exception
+     */
+    public function testCancelViaAPI(): void
     {
         $pirep = $this->createPirep()->toArray();
 
@@ -554,8 +563,10 @@ class PIREPTest extends TestCase
 
     /**
      * When a PIREP is accepted, ensure that the bid is removed
+     *
+     * @throws \Exception
      */
-    public function testPirepBidRemoved()
+    public function testPirepBidRemoved(): void
     {
         $bidSvc = app(BidService::class);
         $flightSvc = app(FlightService::class);
@@ -589,7 +600,7 @@ class PIREPTest extends TestCase
         $this->assertNull($user_bid);
     }
 
-    public function testPirepProgressPercent()
+    public function testPirepProgressPercent(): void
     {
         $this->updateSetting('units.distance', 'km');
 
@@ -620,7 +631,7 @@ class PIREPTest extends TestCase
     /**
      * See that the notifications are properly formatted
      */
-    public function testNotificationFormatting()
+    public function testNotificationFormatting(): void
     {
         $this->updateSetting('units.distance', 'km');
 
@@ -652,7 +663,10 @@ class PIREPTest extends TestCase
         $this->assertEquals('185.2 km', $fields['Distance']);
     }
 
-    public function testDiversionHandler()
+    /**
+     * @throws \Exception
+     */
+    public function testDiversionHandler(): void
     {
         $this->updateSetting('pireps.handle_diversion', true);
         $this->updateSetting('notifications.discord_pirep_diverted', true);
