@@ -17,15 +17,15 @@ use Carbon\Carbon;
 use Illuminate\Support\Facades\Event;
 use Illuminate\Support\Facades\Hash;
 
-class UserTest extends TestCase
+final class UserTest extends TestCase
 {
     /** @var SettingRepository */
-    protected $settingsRepo;
+    protected SettingRepository $settingsRepo;
 
     /** @var UserService */
-    protected $userSvc;
+    protected UserService $userSvc;
 
-    public function setUp(): void
+    protected function setUp(): void
     {
         parent::setUp();
         $this->userSvc = app(UserService::class);
@@ -36,7 +36,7 @@ class UserTest extends TestCase
      * Makes sure that the subfleet/aircraft returned are allowable
      * by the users rank.
      */
-    public function testRankSubfleets()
+    public function testRankSubfleets(): void
     {
         // Add subfleets and aircraft, but also add another
         // set of subfleets
@@ -93,7 +93,7 @@ class UserTest extends TestCase
      *
      * @throws \Exception
      */
-    public function testGetAllAircraft()
+    public function testGetAllAircraft(): void
     {
         $fare_svc = app(FareService::class);
 
