@@ -4,13 +4,15 @@ namespace Tests;
 
 use App\Support\Math;
 use App\Support\Units\Distance;
+use PhpUnitsOfMeasure\Exception\NonNumericValue;
+use PhpUnitsOfMeasure\Exception\NonStringUnitName;
 
-class MathTest extends TestCase
+final class MathTest extends TestCase
 {
     /**
      * Test adding/subtracting a percentage
      */
-    public function testAddPercent()
+    public function testAddPercent(): void
     {
         $tests = [
             ['expected' => 112, 'fn' => Math::getPercent(100, 112)],
@@ -30,7 +32,11 @@ class MathTest extends TestCase
         }
     }
 
-    public function testDistanceMeasurement()
+    /**
+     * @throws NonNumericValue
+     * @throws NonStringUnitName
+     */
+    public function testDistanceMeasurement(): void
     {
         $dist = new Distance(1, 'mi');
         $this->assertEqualsWithDelta(1609.34, $dist['m'], 0.1);
